+++ router/leaguesAPI.router.js
@@ -1,5 +1,5 @@
 const express = require("express");
-const updateTeamRank = require("../controller/leaguesApi/updateTeamRank.js");
+const updateTeamRank = require("../controller/leaguesAPI/updateTeamRank.js");
 const wrapAsync = require("../util/wrapAsync");
 
 const router = express.Router();
